Move NativeInputFieldSuppressor to duckchat-api
The :app OnboardingExperimentNativeInputFieldSuppressor contributes to
the NativeInputFieldSuppressor plugin point, but the boundType interface
lived in duckchat-impl. Importing it from :app tripped the
NoImplImportsInAppModule lint rule ("Modules should not import from
:*-impl modules"), failing the Lint CI job.

Move the interface to duckchat-api so consumers can implement it without
an impl import, and keep the @ContributesPluginPoint trigger in
duckchat-impl (where Anvil is allowed). This mirrors the existing
boundType-in-api / trigger-in-impl pattern (e.g. HatchInteractionsPlugin).
